Add the Line SpellShape.
The origin of this shape is not the center like the current shapes. Behaviorially it is similar to the circle shape. Think of the length as the radius of a circle. For now set the tile_size to 1 by 1 so it always snaps to the the center of the tile.